Move compute-next-step to its own folder
