+++ packages/@coorpacademy-progression-engine/src/compute-next-step/index.js
@@ -15,8 +15,8 @@ import last from 'lodash/fp/last';
 import filter from 'lodash/fp/filter';
 import includes from 'lodash/fp/includes';
 import intersection from 'lodash/fp/intersection';
-import type {State, Slide, Content, Engine, Config} from './types';
-import getConfig from './config';
+import type {State, Slide, Content, Engine, Config} from '../types';
+import getConfig from '../config';
 
 const isAlive = (state: State): boolean => state.lives > 0;
 const hasRemainingLifeRequests = (state: State): boolean => state.remainingLifeRequests > 0;
