+++ packages/@coorpacademy-progression-engine/src/compute-next-step/test/compute-next-step.js
@@ -6,8 +6,8 @@ import merge from 'lodash/fp/merge';
 import pipe from 'lodash/fp/pipe';
 import find from 'lodash/fp/find';
 import omit from 'lodash/fp/omit';
-import type {State} from '../types';
-import computeNextStep from '../compute-next-step';
+import type {State} from '../../types';
+import computeNextStep from '..';
 import allSlides from './fixtures/slides';
 import {stateBeforeGettingNextContent} from './fixtures/states';
